refactor: TimeParser와 TimePolicy로 분리
