+++ src/test/java/domain/policy/TimePolicyTest.java
@@ -1,10 +1,12 @@
 package domain.policy;
 
-import view.Policy.TimePolicy;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import java.time.LocalTime;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -18,35 +20,10 @@ public class TimePolicyTest {
     }
 
     @ParameterizedTest
-    @DisplayName("시의 범위를 벗어나면 예외가 발생한다")
-    @ValueSource(strings = {"-1", "24", "a"})
-    void validateHourTest(String hour){
-        assertThatThrownBy(() -> timePolicy.validateHour(hour))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @ParameterizedTest
-    @DisplayName("분의 범위를 벗어나면 예외가 발생한다")
-    @ValueSource(strings = {"60", "-1", "a"})
-    void validateMinuteTest(String minute) {
-        assertThatThrownBy(() -> timePolicy.validateMinute(minute))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @ParameterizedTest
-    @DisplayName("시간 입력 형식이 다르면 예외가 발생한다")
-    @ValueSource(strings = {"12/34", "12:12:12"})
-    void validateTimeFormatTest(String time) {
-        assertThatThrownBy(() -> timePolicy.validateTimeFormat(time))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @ParameterizedTest
     @DisplayName("캠퍼스 운영 시간이 아니면 예외가 발생한다")
-    @ValueSource(strings = {"07:59", "23:01"})
-    void validateOperatingTimeTest(String time) {
+    @CsvSource({"07:59", "23:01"})
+    void validateOperatingTimeTest(LocalTime time) {
         assertThatThrownBy(() -> timePolicy.validateOperatingTime(time))
                 .isInstanceOf(IllegalArgumentException.class);
     }
-
 }
